refactor(doctor): extract allowlist store recovery helpers
